.nextmv/release/requirements.txt
boto3>=1.34.33
pyyaml>=6.0.1
ruff>=0.1.7
nextmv==v0.10.3
